# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller._ResponsePostList:
    properties:
      code:
        description: 业务响应状态码
        type: string
      data:
        description: 数据
        items:
          $ref: '#/definitions/models.ApiPostDetail'
        type: array
      message:
        description: 提示信息
        type: string
    type: object
  models.ApiPostDetail:
    properties:
      author_id:
        type: integer
      author_name:
        type: string
      community_id:
        type: integer
      content:
        type: string
      create_time:
        type: string
      id:
        type: integer
      introduction:
        type: string
      name:
        type: string
      status:
        type: integer
      title:
        type: string
      vote_num:
        type: integer
    required:
    - community_id
    - content
    - title
    type: object
  models.ParamPostList:
    properties:
      community_id:
        description: 社区ID　可以为空
        type: integer
      order:
        description: 排序依据
        example: score
        type: string
      page:
        description: 页码
        example: 1
        type: integer
      size:
        description: 每页数据量
        example: 10
        type: integer
    type: object
  models.Post:
    properties:
      author_id:
        type: integer
      community_id:
        type: integer
      content:
        type: string
      create_time:
        type: string
      id:
        type: integer
      status:
        type: integer
      title:
        type: string
    required:
    - community_id
    - content
    - title
    type: object
host: localhost:8081
info:
  contact:
    email: 1715925630@nyist.edu.cn
    name: doudoudoubi1220
  license:
    name: Apache 2.0
    url: http://www.apache.org/license/LICENSE-2.0html
  title: bluebell
  version: "1.0"
paths:
  /post:
    post:
      consumes:
      - application/json
      description: 根据用户输入的数据创建一个帖子
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostList'
      security:
      - ApiKeyAuth: []
      summary: 创建帖子接口
      tags:
      - 帖子相关接口
  /post/{id}:
    get:
      consumes:
      - application/json
      description: 根据传入的postid查询帖子的详细信息
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostList'
      security:
      - ApiKeyAuth: []
      summary: 获取帖子详情的接口
      tags:
      - 帖子相关接口
  /post2:
    get:
      consumes:
      - application/json
      description: 可按社区按时间或分数排序查询帖子列表接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 查询参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.ParamPostList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostList'
      security:
      - ApiKeyAuth: []
      summary: 升级版帖子列表接口
      tags:
      - 帖子相关接口
  /posts:
    get:
      consumes:
      - application/json
      description: 获取所有帖子列表，根据传递的参数进行分页，按照发部顺序进行排序
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostList'
      security:
      - ApiKeyAuth: []
      summary: 获取帖子列表的接口
      tags:
      - 帖子相关接口
swagger: "2.0"